backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "FastAPI backend for Todo Web App"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "sqlmodel>=0.0.16",
    "aiosqlite>=0.17.0",
    "python-jose[cryptography]>=3.3.0",

    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.9",
    "bcrypt",
]
